- added a bunch of code that helps to track the improper use of array indices, using the c++ strong type checking. - eventually we will probably port the code completely over to c++, so I added several things ("by_reference" and some declarations of some variables as const) that will help optimize the code in c++. - I am also uploading the Fluid Studios memory manager (which only works in c++) that I have previously used to track memory leaks

+++ game/extensions/SDL_extensions.c
@@ -693,7 +693,7 @@ SDL_bool SDLX_ExpandFormat( SDL_PixelFormat * pformat )
 {
     // use the basic screen format to create a surface with proper alpha support
 
-    int i;
+    Uint8 i;
 
     if ( NULL == pformat ) return SDL_FALSE;
